requirements-minimal.txt
fastapi==0.110.1
uvicorn==0.25.0
motor==3.3.1
python-dotenv==1.2.1
pydantic==2.12.5
pymongo==4.5.0
dnspython==2.8.0
python-dateutil==2.9.0.post0
